+++ src/CG.Blazor/ServiceCollectionExtensions.cs
@@ -2,7 +2,6 @@
 using CG.Validations;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 
 namespace CG.Blazor
 {
@@ -41,6 +40,67 @@ namespace CG.Blazor
 
             // Add a token provider service.
             serviceCollection.Add<TokenProvider>(serviceLifetime);
+
+            // Return the service collection.
+            return serviceCollection;
+        }
+
+        // *******************************************************************
+
+        /// <summary>
+        /// This method registers a <see cref="IStateService"/> service.
+        /// </summary>
+        /// <param name="serviceCollection">The service collection to use 
+        /// for the operation.</param>
+        /// <param name="configuration">The configuration to use for the 
+        /// operation.</param>
+        /// <param name="serviceLifetime">The service lifetime to use for
+        /// the operation.</param>
+        /// <returns>The value of the <paramref name="serviceCollection"/>
+        /// parameter, for chaining calls together.</returns>
+        public static IServiceCollection AddStateService(
+            this IServiceCollection serviceCollection,
+            IConfiguration configuration,
+            ServiceLifetime serviceLifetime = ServiceLifetime.Scoped
+            )
+        {
+            // Validate the parameters before attempting to use them.
+            Guard.Instance().ThrowIfNull(serviceCollection, nameof(serviceCollection))
+                .ThrowIfNull(configuration, nameof(configuration));
+
+            // Add a state service.
+            serviceCollection.Add<IStateService, StateService>(serviceLifetime);
+
+            // Return the service collection.
+            return serviceCollection;
+        }
+
+        // *******************************************************************
+
+        /// <summary>
+        /// This method registers a <see cref="IStateService"/> service.
+        /// </summary>
+        /// <typeparam name="T">The type of state service to add.</typeparam>
+        /// <param name="serviceCollection">The service collection to use 
+        /// for the operation.</param>
+        /// <param name="configuration">The configuration to use for the 
+        /// operation.</param>
+        /// <param name="serviceLifetime">The service lifetime to use for
+        /// the operation.</param>
+        /// <returns>The value of the <paramref name="serviceCollection"/>
+        /// parameter, for chaining calls together.</returns>
+        public static IServiceCollection AddStateService<T>(
+            this IServiceCollection serviceCollection,
+            IConfiguration configuration,
+            ServiceLifetime serviceLifetime = ServiceLifetime.Scoped
+            ) where T : class, IStateService
+        {
+            // Validate the parameters before attempting to use them.
+            Guard.Instance().ThrowIfNull(serviceCollection, nameof(serviceCollection))
+                .ThrowIfNull(configuration, nameof(configuration));
+
+            // Add a state service.
+            serviceCollection.Add<IStateService, T>(serviceLifetime);
 
             // Return the service collection.
             return serviceCollection;